fix: address Copilot review feedback on PR #70
Five fixes from the Copilot review:

1. Bare ValueError on BudgetManager validation violated AGENTS.md
   ("never raise bare ValueError to callers"). Replaced with a new
   BudgetConfigError(AgentKernelError); updated tests.

2. firewall/budgets.py exceeded the ≤300 line guideline. Split into:
   - budgets.py (28 lines, original Budgets dataclass only)
   - token_counting.py (41 lines, TokenCounter + default_token_counter)
   - budget_manager.py (275 lines, BudgetManager + helpers)
   Public imports unchanged; everything re-exported via firewall/__init__.

3. invoke() did not mirror the Firewall's admin-only raw gate. A
   non-admin requesting raw kept effective_mode == "raw", which made
   the kernel skip handle creation even though the Firewall would then
   downgrade to summary — yielding a summary frame without a handle.
   The kernel now applies the same raw → summary downgrade before the
   budget escalation and handle-creation decision. Added a regression
   test covering the case.

4. A Firewall exception after a budget reservation permanently leaked
   the reserved tokens. Wrapped the firewall transform + reconciliation
   in try/finally that releases the reservation if record_usage never
   ran. Added a regression test using a stub failing Firewall.

5. Updated firewall tests to assert BudgetConfigError instead of
   ValueError, and verified BudgetConfigError is a subclass of
   AgentKernelError.

make ci: lint clean, mypy clean, 403 tests pass (was 400; +3 from
this change), all examples run.
